requirements.txt
livekit-agents==1.2.6
livekit-plugins-openai==1.2.6
livekit-plugins-deepgram==1.2.6
livekit-plugins-silero==1.2.6
livekit-plugins-cartesia==1.2.6
python-dotenv==1.0.1
requests==2.32.3
